Stroke expansion
Trace stroke can be now expanded up to any size. In addition, the graph
weights are now updated according to updates of pixel intensity. This
is an important feature to make traces avoid themselves, but it makes
the algorithm slower. Comment the line 138 Im2Graph ->
UpdateWeights(image,vtkG,Buffer); to turn off this feature.